refactor data management
