+++ workbase/src/renderer/components/shared/StoresActions.js
@@ -13,6 +13,7 @@ export const ADD_TYPE = 'add-type';
 
 // Data Management store actions
 export const RUN_CURRENT_QUERY = 'run-current-query';
+export const EXPLAIN_CONCEPT = 'explain-concept';
 export const TOGGLE_LABEL = 'toggle-label';
 export const TOGGLE_COLOUR = 'toggle-colour';
